.htaccess, +.gitignore

+++ controllers/AuthassignmentController.php
@@ -6,7 +6,9 @@ use app\objects\ViewModels\AuthAssignmentCreateView;
 use Yii;
 use app\models\AuthAssignment;
 use app\models\search\AuthAssignmentSearch;
+use yii\filters\AccessControl;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -27,6 +29,15 @@ class AuthassignmentController extends Controller
                     'delete' => ['POST'],
                 ],
             ],
+            'access' => [
+                'class' => AccessControl::class,
+                'rules' => [
+                    [
+                        'allow' => true,
+                        'roles' => ['@'], // авторизованные пользователи
+                    ],
+                ],
+            ]
         ];
     }
 
@@ -49,33 +60,40 @@ class AuthassignmentController extends Controller
      * @param string $item_name
      * @param string $user_id
      * @return mixed
+     * @throws ForbiddenHttpException
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionView($item_name, $user_id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($item_name, $user_id),
-        ]);
+        if (\Yii::$app->user->can('view')) {
+            return $this->render('view', [
+                'model' => $this->findModel($item_name, $user_id),
+            ]);
+        }
+        throw new ForbiddenHttpException('Недостаточно прав для данного действия');
     }
 
     /**
      * Creates a new AuthAssignment model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
+     * @throws ForbiddenHttpException
      */
     public function actionCreate()
     {
         $model = new AuthAssignment();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'item_name' => $model->item_name, 'user_id' => $model->user_id]);
+        if (\Yii::$app->user->can('create')) {
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'item_name' => $model->item_name, 'user_id' => $model->user_id]);
+            }
+
+            return $this->render('create', [
+                'model' => $model,
+                'viewModel' => new AuthAssignmentCreateView(),
+            ]);
         }
-
-        return $this->render('create', [
-            'model' => $model,
-            'viewModel' => new AuthAssignmentCreateView(),
-
-        ]);
+        throw new ForbiddenHttpException('Недостаточно прав для данного действия');
     }
 
     /**
@@ -84,21 +102,23 @@ class AuthassignmentController extends Controller
      * @param string $item_name
      * @param string $user_id
      * @return mixed
+     * @throws ForbiddenHttpException
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionUpdate($item_name, $user_id)
     {
         $model = $this->findModel($item_name, $user_id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'item_name' => $model->item_name, 'user_id' => $model->user_id]);
+        if (\Yii::$app->user->can('update')) {
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'item_name' => $model->item_name, 'user_id' => $model->user_id]);
+            }
+            return $this->render('update', [
+                'model' => $model,
+                'viewModel' => new AuthAssignmentCreateView(),
+            ]);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-            'viewModel' => new AuthAssignmentCreateView(),
-
-        ]);
+        throw new ForbiddenHttpException('Недостаточно прав для данного действия');
     }
 
     /**
@@ -107,13 +127,18 @@ class AuthassignmentController extends Controller
      * @param string $item_name
      * @param string $user_id
      * @return mixed
+     * @throws ForbiddenHttpException
      * @throws NotFoundHttpException if the model cannot be found
+     * @throws \Throwable
+     * @throws \yii\db\StaleObjectException
      */
     public function actionDelete($item_name, $user_id)
     {
-        $this->findModel($item_name, $user_id)->delete();
-
-        return $this->redirect(['index']);
+        if (\Yii::$app->user->can('delete')) {
+            $this->findModel($item_name, $user_id)->delete();
+            return $this->redirect(['index']);
+        }
+        throw new ForbiddenHttpException('Недостаточно прав для данного действия');
     }
 
     /**
@@ -129,7 +154,6 @@ class AuthassignmentController extends Controller
         if (($model = AuthAssignment::findOne(['item_name' => $item_name, 'user_id' => $user_id])) !== null) {
             return $model;
         }
-
         throw new NotFoundHttpException('The requested page does not exist.');
     }
 }
